kadai-twitterclone

+++ app/Http/Controllers/TasksController.php
@@ -18,13 +18,13 @@ class TasksController extends Controller
         $data = [];
         if (\Auth::check()) {
             $user = \Auth::user();
-            $microposts = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
             $data = [
                 'user' => $user,
                 'tasks' => $tasks,
             ];
-            $data += $this->counts($user);
+            /**$data += $this->counts($user);**/
             return view('tasks.index', $data);
         }else {
             return view('welcome');
@@ -39,7 +39,7 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
+        $task = new task;
 
         return view('tasks.create', [
             'task' => $task,
@@ -59,12 +59,12 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
 
-       
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+       $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            
+        ]);
+        
         return redirect('/');
     }
 
@@ -76,11 +76,14 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-         $task = Task::find($id);
+        $task = \App\task::find($id);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+            'task' => $task,]);}
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -91,11 +94,15 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
+       $task = \App\task::find($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+            'task' => $task,]);}
+        else{
+            return redirect('/');
+        }
+           
     }
 
     /**
@@ -111,7 +118,7 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
-        $task = Task::find($id);
+        $task = task::find($id);
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -127,9 +134,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\task::find($id);
 
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            
+        }
+    
         return redirect('/');
     }
 }
